move cli to another module
