+++ src/lib.rs
@@ -2,3 +2,4 @@ pub mod hash;
 pub mod merkle_error;
 pub mod merkle_tree;
 pub mod proof_element;
+pub mod cli;
